Add solution for Week 0, Project 0: Tic-Tac-Toe
Solution for Week 0: Search, Project 0: Tic-Tac-Toe in CS50's
Introduction to Artificial Intelligence with Python. Focusing on
an adversarial search algorithm for the game of Tic-Tac-Toe. Using the
Minimax algorithm and employing Alpha-Beta Pruning for optimisation, we
develop an AI player that will consistently make the optimal move for
every state, and achieve a draw or a win against the player.