Add solution for Week 1, Project 1: Minesweeper
Solution for Week 1, Project 1: Knights, this problem set focuses on
building your own Minesweeper AI. The task was to build a
knowledge-based agent to make decisions and inferences based on the
knowledge it builds up after every move. The AI will always select a
cell it knows to be safe, and always avoid a cell it knows to be a mine,
but may still make mistakes when selecting a random move or making an
informed guess. A challenging task to optimise as much as possible!