Add solution for Week 2, Project 2: Page Rank
Solution for Week 2, Project 2: Page Rank, as the name suggests, this
problem dealt with the PageRank algorithm developed by Google. It's
purpose being to assign a rank to pages based on their perceived
'importance'. I used two strategies to calculate these rankings. The
Random Surfer Model (interpreted as a Markov Chain), and the Iterative
Algorithm, which uses a recursive mathematical expression to calculate
Page Ranks until the values eventually converge.