Add solution for Week 2, Project 2: Heredity
Solution for Week 2, Project 2: Heredity. This problem involved creating
an AI to assess the probability of a person inheriting a particular gene
and exhibiting a particular trait. I dealt with probabilities at a level
I hadn't experienced before and came out of this project learning quite
a lot. Inevitably, due to the branching nature of each possibility,
and dependence on a child's ancestry, I had a tougher time than expected
with the calculations, but after restarting and refactoring several
times, I have a set of functions I am quite proud of.