Add solution for Week 4, Project 4: Nim
Solution for Week 4, Project 4: Nim. This project involved implementing
an AI that teaches itself to play Nim through reinforcement learning.
Optionally using the ε-Greedy to determine the AI's action in the game.
The AI uses Q-Learning to learn a reward value for each (state, action)
pair, and a learning rate (α) to weigh the value of new information
against information it already has. Short and simple, and disheartening
to play against when it's been trained on 1000 games!