Add solution for Week 5, Project 5: Traffic
Solution for Week 5, Project 5: Traffic. This project was an exercise in
computer vision and machine learning. Involving the creation of a
Convolutional Neural Network trained on a dataset of road signs to
identify the category (and therefore, meaning) of images of those signs.

Using TensorFlow and OpenCV2, the images are first converted into
3-channel (RGB) multi-dimensional arrays. Then split into a training and
test set, and fed through a CNN of our own making. The project
encouraged lots of testing of various different parameters, layers, and
general architecture, to arrive at an acceptable level of accuracy with
an inexpensive computational cost. I took a lot from this and plan to
work more with TensorFlow in my future projects.