Add solution for Week 6, Project 6: Parser
Solution for Week 6, Project 6: Parser. This one involved natural
language processing to parse sentences and extract noun phrases. Had to
brush up on my linguistics to get everything right, but it was very
rewarding to see everything processed properly only after writing a few
simple rules about ho to put together sentences. It was interesting to
take on the role of English teacher to my computer, and go through
nouns, verbs, conjugations, etc. all over again.

Unfortunately did not pass the final check according to check50, however
this is a well documented error and the project can be submitted and
'passed' otherwise.